# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Anime Calendar API
  version: 1.0.0
  description: |
    REST API for Anime Calendar — create and manage personal anime/manga calendars
    sourced from AniList, exportable as iCalendar (.ics) feeds.

    ## Authentication
    Most endpoints require a Bearer JWT obtained via `/login`, `/register`, or `/auth/google`.
    Include it as: `Authorization: Bearer <token>`

    ## Error responses
    All errors return a JSON body: `{"error": "<message>"}`.
    The HTTP status code indicates the error category.
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------------------------------------------------------
  # Reusable response components
  # -------------------------------------------------------------------------
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
...
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid request data
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Requested resource not found
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 128
          example: S3cur3P@ssword
    LoginResponse:
      type: object
      required: [token, username]
      properties:
        token:
          type: string
          description: JWT bearer token (24h expiry)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        username:
          type: string
          example: animefan42
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          maxLength: 50
          example: animefan42
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          maxLength: 128
          description: Must meet strength requirements (mixed case, digits, symbols)
          example: S3cur3P@ssword
...
    TokenVerifyRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: JWT to validate
...
security:
  - bearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /register:
    post:
      tags: [Auth]
      summary: Register a new account
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful — returns JWT and username
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/verify:
    post:
      tags: [Auth]
      summary: Verify a JWT is still valid
      operationId: verifyToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenVerifyRequest'
      responses:
        '200':
          description: Token is valid
        '401':
          $ref: '#/components/responses/Unauthorized'
...